fixed contraint issue

+++ database/migrations/2021_07_03_083809_create_review_fasilitas_table.php
@@ -15,7 +15,7 @@ class CreateReviewFasilitasTable extends Migration
     {
         Schema::create('review_fasilitas', function (Blueprint $table) {
             $table->id('id_reviewFasilitas');
-            $table->integer('id_penginapan');
+            $table->bigInteger('id_penginapan');
             $table->longText('review');
             $table->foreign('id_penginapan')->references('id_penginapan')->on('penginapans')->onDelete('cascade');
         });
