+++ database/migrations/2021_07_03_083821_create_review_kesehatans_table.php
@@ -15,7 +15,7 @@ class CreateReviewKesehatansTable extends Migration
     {
         Schema::create('review_kesehatans', function (Blueprint $table) {
             $table->id('id_reviewKesehatan');
-            $table->integer('id_penginapan');
+            $table->bigInteger('id_penginapan');
             $table->longText('review');
             $table->foreign('id_penginapan')->references('id_penginapan')->on('penginapans')->onDelete('cascade');
         });
